+++ scripts/run_example_tp.py
@@ -162,7 +162,61 @@ if __name__ == "__main__":
             do_random=args.random,
         )
     elif args.dataset == "pile":
-        from utils import get_prompts, get_prompt_token_ids, get_pile_prompts
+
+        def reset_seed(seed=42):
+            import torch
+            import random
+            import numpy as np
+
+            torch.manual_seed(seed)
+            np.random.seed(seed)
+            random.seed(seed)
+
+        def get_prompt_token_ids(model_path, prompts, max_length=1024):
+            from transformers import AutoTokenizer
+
+            tokenizer = AutoTokenizer.from_pretrained(model_path)
+            prompt_token_ids = []
+            for prompt in prompts:
+                tokens = tokenizer(
+                    prompt,
+                    return_tensors="pt",
+                    truncation=True,
+                    max_length=max_length,
+                )
+                if len(tokens.input_ids[0]) < max_length:
+                    continue
+                prompt_token_ids.append([x.item() for x in tokens.input_ids[0]])
+            return prompt_token_ids
+
+        def get_pile_prompts(model_name, num_samples=512):
+            from datasets import load_dataset
+            from tqdm import tqdm
+            import transformers
+
+            least_tokens = 1024
+            seed = 42
+
+            reset_seed(seed)
+
+            dataset = load_dataset("NeelNanda/pile-10k", split="train")
+            dataset = dataset.shuffle(seed=seed)
+
+            tokenizer = transformers.AutoTokenizer.from_pretrained(
+                model_name, trust_remote_code=True
+            )
+            num_sample = 0
+            samples_lst = []
+            for data in tqdm(dataset):
+                prompt = data["text"]
+                tokens = tokenizer(prompt, return_tensors="pt")
+                if len(tokens.input_ids[0]) < least_tokens:
+                    continue
+                num_sample += 1
+                samples_lst.append(prompt)
+                if num_sample >= num_samples:
+                    break
+            return samples_lst
         least_tokens = args.isl
         num_samples = args.nprompts
         prompts = get_pile_prompts(args.model, num_samples)
